refactor: 상품 카드 컴포넌트 재사용 구조로 개선
